rework compile scripts to remove a lot of steps

+++ tools/sw_compile.sh
@@ -1,8 +1,6 @@
 #!/bin/sh
 
-#/opt/altera16.1_lite/nios2eds/nios2_command_shell.sh
-
-export LM_LICENSE_FILE=
+echo "Software compilation (ORCA)"
 
 cd ../sw/
 
@@ -10,27 +8,53 @@ cd low_level
 make clean
 make
 
+status=$?
+[ $status -eq 0 ] || exit 1
+
+cd ../imu
+make clean
+make
+
+status=$?
+[ $status -eq 0 ] || exit 1
+
 cd ../lidar
 make clean
 make
+
+status=$?
+[ $status -eq 0 ] || exit 1
 
 cd ../pid
 make clean
 make
 
+status=$?
+[ $status -eq 0 ] || exit 1
+
 cd ../odometry
 make clean
 make
 
+status=$?
+[ $status -eq 0 ] || exit 1
+
 cd ../trajectory
 make clean
 make
+
+status=$?
+[ $status -eq 0 ] || exit 1
 
 cd ../../orca/tools
 
 rm low_level.elf.qex
 ./elf2hex.sh ../../sw/low_level/low_level.elf
 cp low_level.elf.qex ../../quartus/system_ll.hex
+
+rm imu.elf.qex
+./elf2hex.sh ../../sw/imu/imu.elf
+cp imu.elf.qex ../../quartus/imu.hex
 
 rm lidar.elf.qex
 ./elf2hex.sh ../../sw/lidar/lidar.elf
@@ -47,14 +71,4 @@ cp odometry.elf.qex ../../quartus/odometry.hex
 rm trajectory.elf.qex
 ./elf2hex.sh ../../sw/trajectory/trajectory.elf
 cp trajectory.elf.qex ../../quartus/trajectory.hex
-
-
-#quartus_cdb --update_mif ../..//quartus/HPSFPGA.qpf
-#quartus_asm ../..//quartus/HPSFPGA.qpf
-#quartus_cpf -c -o bitstream_compression=on ../..//quartus/outputs/HPSFPGA.sof ../..//quartus/outputs/soc_system.rbf
-
-quartus_cdb --rev=DE10 --update_mif ../..//quartus/HPSFPGA.qpf
-quartus_asm --rev=DE10 ../..//quartus/HPSFPGA.qpf
-quartus_cpf -c -o bitstream_compression=on ../..//quartus/outputs/DE10.sof ../..//quartus/outputs/soc_system_de10.rbf
-
